+++ Example/PaymentSheet Example/PaymentSheetUITest/PaymentSheetLPMUITest.swift
@@ -207,6 +207,25 @@ class PaymentSheetStandardLPMUIOneTests: PaymentSheetStandardLPMUICase {
         XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 15.0))
     }
 
+    func testCryptoPaymentMethod() throws {
+        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
+        settings.layout = .horizontal
+        settings.currency = .usd
+        settings.merchantCountryCode = .US
+        loadPlayground(app, settings)
+        app.buttons["Present PaymentSheet"].tap()
+
+        // Select Crypto
+        tapPaymentMethod("Crypto")
+
+        // Pay
+        app.buttons["Pay $50.99"].waitForExistenceAndTap()
+        webviewAuthorizePaymentButton.waitForExistenceAndTap(timeout: 10)
+
+        // Close the webview, Crypto's test playground is out of scope
+        app.otherElements["TopBrowserBar"].buttons["Close"].waitForExistenceAndTap(timeout: 10)
+    }
+
     func testZipPaymentMethod() throws {
         var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
         settings.layout = .horizontal